Moved node type into node package
