+++ server/ast/discard.go
@@ -17,66 +17,22 @@ package ast
 import (
 	"fmt"
 
-	"github.com/dolthub/go-mysql-server/sql"
+	"github.com/dolthub/doltgresql/server/node"
 	vitess "github.com/dolthub/vitess/go/vt/sqlparser"
 
 	"github.com/dolthub/doltgresql/postgres/parser/sem/tree"
 )
 
 // nodeDiscard handles *tree.Discard nodes.
-func nodeDiscard(node *tree.Discard) (vitess.Statement, error) {
-	if node == nil {
+func nodeDiscard(discard *tree.Discard) (vitess.Statement, error) {
+	if discard == nil {
 		return nil, nil
 	}
-	if node.Mode != tree.DiscardModeAll {
-		return nil, fmt.Errorf("unhandled DISCARD mode: %v", node.Mode)
+	if discard.Mode != tree.DiscardModeAll {
+		return nil, fmt.Errorf("unhandled DISCARD mode: %v", discard.Mode)
 	}
 
 	return vitess.InjectedStatement{
-		Statement: DiscardStatement{},
+		Statement: node.DiscardStatement{},
 	}, nil
-}
-
-// DiscardStatement is just a marker type, since all functionality is handled by the connection handler,
-// rather than the engine. It has to conform to the sql.ExecSourceRel interface to be used in the handler, but this
-// functionality is all unused.
-type DiscardStatement struct{}
-
-var _ vitess.Injectable = DiscardStatement{}
-var _ sql.ExecSourceRel = DiscardStatement{}
-
-func (d DiscardStatement) Resolved() bool {
-	return true
-}
-
-func (d DiscardStatement) String() string {
-	return "DISCARD ALL"
-}
-
-func (d DiscardStatement) Schema() sql.Schema {
-	return nil
-}
-
-func (d DiscardStatement) Children() []sql.Node {
-	return nil
-}
-
-func (d DiscardStatement) WithChildren(children ...sql.Node) (sql.Node, error) {
-	return d, nil
-}
-
-func (d DiscardStatement) CheckPrivileges(ctx *sql.Context, opChecker sql.PrivilegedOperationChecker) bool {
-	return true
-}
-
-func (d DiscardStatement) IsReadOnly() bool {
-	return true
-}
-
-func (d DiscardStatement) RowIter(ctx *sql.Context, r sql.Row) (sql.RowIter, error) {
-	panic("DISCARD ALL should be handled by the connection handler")
-}
-
-func (d DiscardStatement) WithResolvedChildren(children []any) (any, error) {
-	return d, nil
 }
